Add support for Rise Cache
Program flow is that Rise Cache will first be pinged to see if it’s
running. Then a request is made to Rise Storage to return the files. If
Rise Cache is running, the URLs will be altered so that they come from
Rise Cache; otherwise, the URLs are passed as is from Storage.